update PaginationResult property to count from int to long

+++ Backend/Common/Pagination/PaginationResult.cs
@@ -1,12 +1,12 @@
 ﻿namespace Common.Pagination
 {
     public class PaginationResult<TEntity>
-        (int pageNumber, int pageSize, int count, IEnumerable<TEntity> data) 
+        (int pageNumber, int pageSize, long count, IEnumerable<TEntity> data) 
         where TEntity : class
     {
         public int PageNumber { get; } = pageNumber;
         public int PageSize { get; } = pageSize;
-        public int Count { get; } = count;
+        public long Count { get; } = count;
         public IEnumerable<TEntity> Data { get; } = data;
     }
 }
